requirements.txt
asgiref==3.4.1
bottlenose==1.1.8
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
idna==3.2
lxml==4.6.3
Pillow==8.3.1
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
